admin home page working with web sockets and made changes in the database design where  the attendance table was changed to incorporate composite key (lecture id , student id)

+++ src/java/hod/controllers/HodHome.java
@@ -33,7 +33,7 @@ public class HodHome extends Controller {
 
         String departmentString = req.getParameter("departmentId");
 
-        int departmentId;
+        long departmentId;
         if (departmentString != null) {
             departmentId = Integer.valueOf(departmentString);
         } else {
